Hyphenation support.
This patch adds optional support for hyphenation using discretionary
hyphens, via the -h option, using TeX::Hyphen.  If TeX::Hyphen is not
installed, -h doesn't work, but everything else should.

(You might also not want to use -h if your document will be shown on the
Kindle Keyboard or older, since the original MOBI reader, all that is
available on hardware of that vintage, renders discretionary hyphens as
hard hyphens!)